+++ packages/react-magma-dom/src/components/ButtonGroup/ButtonGroup.stories.tsx
@@ -119,6 +119,11 @@ export default {
         type: 'boolean',
       },
     },
+    isList: {
+      control: {
+        type: 'boolean',
+      },
+    },
   },
 } as Meta;