+++ packages/react-magma-dom/src/components/FormFieldContainer/FormFieldContainer.tsx
@@ -2,9 +2,11 @@ import * as React from 'react';
 
 import styled from '@emotion/styled';
 
+import useDeviceDetect from '../../hooks/useDeviceDetect';
 import { InverseContext, useIsInverse } from '../../inverse';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { descriptionSuffix, labelSuffix } from '../../utils';
+import { Announce } from '../Announce';
 import { CharacterCounter } from '../CharacterCounter';
 import { InputMessage } from '../Input/InputMessage';
 import { InputIconPosition, InputSize } from '../InputBase';
@@ -187,6 +189,7 @@ export const FormFieldContainer = React.forwardRef<
   } = props;
   const theme = React.useContext(ThemeContext);
   const isInverse = useIsInverse(isInverseProp);
+  const { isWindows, isChrome } = useDeviceDetect();
 
   const countProps = maxCount || maxLength;
   const counterDescriptionId =
@@ -259,6 +262,16 @@ export const FormFieldContainer = React.forwardRef<
                 <>{errorMessage ? errorMessage : helperMessage}</>
               )}
             </InputMessage>
+          )}
+
+          {!(isWindows && isChrome) && (
+            <VisuallyHidden>
+              <Announce>
+                {(errorMessage || helperMessage) && (
+                  <>{errorMessage ? errorMessage : helperMessage}</>
+                )}
+              </Announce>
+            </VisuallyHidden>
           )}
         </InputPositionWrapper>
       </StyledFormFieldContainer>