+++ packages/react-magma-dom/src/i18n/interface.ts
@@ -357,4 +357,8 @@ export interface I18nInterface {
     warning: string;
     danger: string;
   };
+  expansionState: {
+    expanded: string;
+    collapsed: string;
+  };
 }